requirements.txt
Django==3.1.4
psycopg2-binary==2.8.5
djangorestframework==3.12.1
pyyaml==5.3.1

django-filter==2.4.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.5.0
ujson==4.0.1

fastjsonschema==2.14.5
flake8